new option to concatenate files

+++ MC/condor/makeCondor.py
@@ -7,6 +7,7 @@ def getArgs() :
     parser.add_argument("--nickName",default='MCpileup',help="Data set nick name.") 
     parser.add_argument("-m","--mode",default='anaXRD',help="Mode (script to run).")
     parser.add_argument("-y","--year",default=2017,type=str,help="Data taking period, 2016, 2017 or 2018")
+    parser.add_argument("-c","--concatenate",default=4,type=int,help="On how many files to run on each job")
     return parser.parse_args()
 
 def beginBatchScript(baseFileName) :
@@ -47,23 +48,39 @@ if len(files) < 1 :
     exit()
 
 scriptList = [] 
-for nFile, file in enumerate(files) :
-    print("nFile={0:d} file[:80]={1:s}".format(nFile,file[:80]))
+file=[]
+dataset=[]
+
+mjobs=args.concatenate
+
+
+for nFiles, file in enumerate(files) :
+    fileName=getFileName(file)
+    dataset.append(fileName)
+
+
+for nFile in range(0, len(dataset),mjobs) :
+    #print("nFile={0:d} file[:80]={1:s}".format(nFile,file[:80]))
 
     scriptName = "{0:s}_{1:03d}.csh".format(args.nickName,nFile+1)
     print("scriptName={0:s}".format(scriptName))
     outLines = beginBatchScript(scriptName)
-    fileName = getFileName(file)
 
-    outFileName = "{0:s}_{1:03d}.root".format(args.nickName,nFile+1)
-    outLines.append("xrdcp root://cms-xrd-global.cern.ch/{0:s} inFile.root\n".format(fileName)) 
     outLines.append("tar -zxvf SFs.tar.gz\n")
     outLines.append("cp MCsamples_*csv MCsamples.csv\n")
-    outLines.append("python ZH.py -f inFile.root -o {0:s} --nickName {1:s}\n".format(outFileName,args.nickName))
-    outLines.append("mv inFile.csv {0:s}\n".format(outFileName.replace(".root",".csv")))
-    outLines.append("rm inFile.root\n")
+        #print '===========================', fileloop, nFile, file
+
+    fileName = getFileName(file)
+    if nFile+mjobs <= len(dataset)  :
+        for j in range(0,mjobs) :
+            fileloop=dataset[nFile:nFile+mjobs][j]
+            outLines.append("xrdcp root://cms-xrd-global.cern.ch/{0:s} inFile.root\n".format(fileloop)) 
+            outFileName = "{0:s}_{1:03d}.root".format(args.nickName,nFile+j)
+            outLines.append("python ZH.py -f inFile.root -o {0:s} --nickName {1:s} -y {2:s} -l AZH -w 1\n".format(outFileName,args.nickName, args.year))
+            outLines.append("rm inFile.root\n")
+
+
     outLines.append("rm *.pyc\nrm *.so\nrm *.pcm\nrm *cc.d\n")
-    
     print("Writing out file = {0:s}".format(scriptName))
     open(scriptName,'w').writelines(outLines)
     scriptList.append(scriptName)
@@ -95,8 +112,8 @@ for file in scriptList :
     outLines.append('Error = {0:s}.err\n'.format(base))
     outLines.append('Log = {0:s}.log\n'.format(base))
     print("dir={0:s}".format(dir))
-    outLines.append('transfer_input_files = {0:s}ZH.py, {0:s}MC_2017.root, {0:s}data_pileup_2017.root, {0:s}MCsamples_{1:s}.csv, {0:s}ScaleFactor.py, {0:s}SFs.tar.gz, '.format(dir,args.year))
-    outLines.append('{0:s}Cert_*txt, '.format(dirData))
+    outLines.append('transfer_input_files = {0:s}ZH.py, {0:s}MC_{1:s}.root, {0:s}data_pileup_{1:s}.root, {0:s}MCsamples_{1:s}.csv, {0:s}ScaleFactor.py, {0:s}SFs.tar.gz, '.format(dir,args.year))
+    #outLines.append('{0:s}*txt, '.format(dirData))
     outLines.append('{0:s}tauFun.py, {0:s}generalFunctions.py, {0:s}outTuple.py,'.format(funcsDir))
     outLines.append('{0:s}FastMTT.h, {0:s}MeasuredTauLepton.h, {0:s}svFitAuxFunctions.h,'.format(SVFitDir)) 
     outLines.append('{0:s}FastMTT.cc, {0:s}MeasuredTauLepton.cc, {0:s}svFitAuxFunctions.cc\n'.format(SVFitDir))
